refactor(datahub): 阶段 1-2 完成 - helpers 层重构和 domains 层统一
## 阶段 1：文档清理
- 创建 helpers/README.md 说明纯函数工具用途
- 更新 stores/README.md 明确基础设施定位
- 更新 packages/datahub/README.md 架构说明
- 删除 utils/README.md

## 阶段 2：Domains 层统一
- 重命名 accessors/ → helpers/（只保留纯函数工具）
  - adjustment.py: 复权计算（QFQ/HFQ）
  - pit.py: PIT 查询辅助函数
  - 移除 enrichment.py（已搬迁到 ditto-port）
  - 移除 instrument_accessor.py（功能迁移到 MetadataService）

- 重命名 Features 域 IndicatorStore
  - indicator_store.py → technical_indicator_store.py
  - indicator_metadata_store.py → technical_indicator_metadata_store.py
  - 保持向后兼容别名

- 迁移 UniverseStore 到 domains/metadata/universe/
  - 从 stores/ 迁移到 domains 层
  - 更新所有导入引用

- 扩展 MetadataService 添加批量注册方法
  - register_securities_batch(): 批量注册证券
  - resolve_or_create_batch(): 批量解析或创建证券

- 更新依赖注入配置
  - hub.py: 移除 InstrumentsAccessor，securities → metadata
  - Provider 配置: 移除 securities provider
  - data_writer.py: 使用 MetadataService 新 API

## 架构变更
- helpers/ : 纯函数工具（复权、PIT）
- domains/ : 业务域层（Service + Store）
- stores/ : 基础设施（SQLite、Parquet）

## 验证
- 基于类型检查: 0 errors
- 代码检查: ruff All checks passed
