+++ packages/datahub/src/ditto_datahub/domains/features/technical/__init__.py
@@ -1,9 +1,11 @@
 """Technical indicators subdomain - 技术指标子域."""
 
-from ditto_datahub.domains.features.technical.indicator_metadata_store import (
-    IndicatorMetadataStore,
+from ditto_datahub.domains.features.technical.technical_indicator_metadata_store import (  # noqa: E501
+    TechnicalIndicatorMetadataStore as IndicatorMetadataStore,
 )
-from ditto_datahub.domains.features.technical.indicator_store import IndicatorStore
+from ditto_datahub.domains.features.technical.technical_indicator_store import (
+    TechnicalIndicatorStore as IndicatorStore,
+)
 
 __all__ = [
     "IndicatorMetadataStore",
