+++ packages/datahub/src/ditto_datahub/domains/metadata/instrument/instrument_store.py
@@ -19,9 +19,6 @@ import polars as pl
 from ditto_foundation import M, logger, traced
 from ditto_foundation.cache import DataCache
 
-from ditto_datahub.accessors.internal.enrichment import (
-    enrich_with_symbol as enrich_with_symbol_fn,
-)
 from ditto_datahub.domains.metadata.instrument.models import InstrumentRegistration
 from ditto_datahub.stores.sqlite_client import SQLiteClient
 
@@ -508,8 +505,8 @@ class InstrumentStore:
             }
         )
 
-        # 使用纯函数进行数据增强
-        return enrich_with_symbol_fn(df, symbol_df)
+        # 内联数据增强：join symbol 数据
+        return df.join(symbol_df, on="sid", how="left")
 
     def register(self, sid: int, registration: InstrumentRegistration) -> int:
         """
