+++ packages/datahub/src/ditto_datahub/domains/metadata/universe/__init__.py
@@ -1,0 +1,5 @@
+"""Universe subdomain - 标的池子域."""
+
+from ditto_datahub.domains.metadata.universe.universe_store import UniverseStore
+
+__all__ = ["UniverseStore"]
