+++ packages/datahub/src/ditto_datahub/helpers/__init__.py
@@ -1,0 +1,15 @@
+"""
+Helpers - 纯函数工具模块。
+
+提供无副作用的纯函数工具，用于复权计算和 PIT 查询支持。
+"""
+
+from ditto_datahub.helpers.adjustment import apply_hfq_adj, apply_qfq_adj
+from ditto_datahub.helpers.pit import filter_by_knowledge_date, parse_asof_date
+
+__all__ = [
+    "apply_hfq_adj",
+    "apply_qfq_adj",
+    "filter_by_knowledge_date",
+    "parse_asof_date",
+]
